+++ model/pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     
     <parent>
-        <artifactId>phonebook</artifactId>
-        <groupId>com.kozubal</groupId>
+        <artifactId>poster</artifactId>
+        <groupId>org.pti</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
