+++ pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.kozubal</groupId>
-    <artifactId>phonebook</artifactId>
+    <groupId>org.pti</groupId>
+    <artifactId>poster</artifactId>
     <version>1.0-SNAPSHOT</version>
     <modules>
         <module>rest</module>
@@ -17,6 +17,8 @@
         <jersey.version>2.15</jersey.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <jersey.framework.version>2.15</jersey.framework.version>
+        <couchbase.version>3.0.0-beta2</couchbase.version>
+        <lombok.version>1.14.4</lombok.version>
     </properties>
 
     <packaging>pom</packaging>
@@ -31,11 +33,17 @@
                 <scope>import</scope>
             </dependency>
 
+            <dependency>
+                <groupId>org.mongodb</groupId>
+                <artifactId>mongo-java-driver</artifactId>
+                <version>${couchbase.version}</version>
+            </dependency>
+
             <!-- for integration module -->
             <dependency>
                 <groupId>org.projectlombok</groupId>
                 <artifactId>lombok</artifactId>
-                <version>1.14.4</version>
+                <version>${lombok.version}</version>
             </dependency>
 
             <dependency>
