+++ rest/src/main/webapp/WEB-INF/web.xml
@@ -10,7 +10,7 @@
         <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
         <init-param>
             <param-name>jersey.config.server.provider.packages</param-name>
-            <param-value>com.phonebook</param-value>
+            <param-value>com.pti</param-value>
         </init-param>
     </servlet>
