+++ views/pom.xml
@@ -3,8 +3,8 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>phonebook</artifactId>
-        <groupId>com.kozubal</groupId>
+        <artifactId>poster</artifactId>
+        <groupId>org.pti</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
@@ -33,7 +33,7 @@
         </dependency>
 
         <dependency>
-            <groupId>com.kozubal</groupId>
+            <groupId>org.pti</groupId>
             <artifactId>model</artifactId>
             <version>${project.version}</version>
         </dependency>
